+++ libft/ft_itoa_base.c
@@ -6,13 +6,13 @@
 /*   By: aschukin <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/07 14:31:42 by aschukin          #+#    #+#             */
-/*   Updated: 2018/02/15 13:12:56 by aschukin         ###   ########.fr       */
+/*   Updated: 2018/03/24 17:45:53 by aschukin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static int	ft_count_base(int num, int base)
+static int	ft_count_base(long long num, int base)
 {
 	int	i;
 
@@ -28,9 +28,8 @@ static int	ft_count_base(int num, int base)
 }
 
 
-char	*ft_itoa_base(int num, int base)
+char	*ft_itoa_base(long long num, int base)
 {
-	// cast int in long long?
 	char	*str;
 	int		count;
 	int		i;
